Work for Friday Sep 1 2023
I got the kinetics dataset to work on a small subsample of the data and
figured out conclusively that I do not have compute optical flow for the
das images (or use optical flow at all) (this is what the self.get_flow
parameter of the Kinetics dataset is doing).

Now it remains to get the model training on the small subset of data you
have downloaded.
By doing that, you shouldhave everything you need to refactor a version
of the dataset/loader that works with dsa images.