Work for labor day: converting niftis to more useful forms
After much ado about nothing (I spent too long trying to figure out
something with filenames that turned out to not be useful at all :()
I have written something that will convert niftis to the mp4 files we
need.
After this, pyav can read them into a container as usual and we can try
and figure out what the data will try and do with it.